Style the account details page
