+++ src/authentication/create-account-details/styles.scss
@@ -1,0 +1,39 @@
+@use '~@zero-tech/zui/styles/theme' as theme;
+
+.create-account-details {
+  color: theme.$color-greyscale-11;
+
+  &__heading {
+    font-weight: 400;
+    font-size: 36px;
+    line-height: 44px;
+    color: theme.$color-greyscale-12;
+    text-shadow: 0px 0px 10px rgba(255, 255, 255, 0.7);
+    margin-top: 80px;
+    margin-bottom: 0px;
+  }
+
+  &__sub-heading {
+    font-weight: 400;
+    font-size: 16px;
+    line-height: 24px;
+    color: theme.$color-greyscale-10;
+    text-align: center;
+    margin-bottom: 32px;
+  }
+
+  &__form {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 14px;
+
+    & > * {
+      width: 280px;
+    }
+  }
+
+  &__button {
+    height: 40px;
+  }
+}
